requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
tk==0.1.0
urllib3==2.3.0
